去掉session

+++ src/main/java/com/mmall/controller/portal/CartController.java
@@ -5,10 +5,14 @@ import com.mmall.common.ResponseCode;
 import com.mmall.common.ServerResponse;
 import com.mmall.pojo.User;
 import com.mmall.service.ICartService;
+import com.mmall.util.CookieUtil;
+import com.mmall.util.JsonUtil;
+import com.mmall.util.RedisPoolUtil;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,6 +20,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import springfox.documentation.annotations.ApiIgnore;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
 /**
@@ -34,9 +39,14 @@ public class CartController {
     @RequestMapping(value = "list.do",method = RequestMethod.GET)
     @ResponseBody
     @ApiOperation("购物车列表")
-    public ServerResponse list(@ApiIgnore HttpSession session) {
-        User user = (User) session.getAttribute(Const.CURRENT_USER);
-        if (user == null) {
+    public ServerResponse list(@ApiIgnore HttpServletRequest request) {
+        String loginToken = CookieUtil.readLoginToken(request);
+        if (StringUtils.isEmpty(loginToken)) {
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
+        }
+        String userJsonStr = RedisPoolUtil.get(loginToken);
+        User user = JsonUtil.string2Obj(userJsonStr,User.class);
+        if (user == null){
             return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
         }
         return iCartService.list(user.getId());
@@ -49,9 +59,14 @@ public class CartController {
             @ApiImplicitParam(value = "产品id",name = "productId",required = true,paramType = "query"),
             @ApiImplicitParam(value = "数量",name = "count",required = true,paramType = "query"),
     })
-    public ServerResponse add(@ApiIgnore HttpSession session,Integer count,Integer productId) {
-        User user = (User) session.getAttribute(Const.CURRENT_USER);
-        if (user == null) {
+    public ServerResponse add(@ApiIgnore HttpServletRequest request,Integer count,Integer productId) {
+        String loginToken = CookieUtil.readLoginToken(request);
+        if (StringUtils.isEmpty(loginToken)) {
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
+        }
+        String userJsonStr = RedisPoolUtil.get(loginToken);
+        User user = JsonUtil.string2Obj(userJsonStr,User.class);
+        if (user == null){
             return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
         }
         return iCartService.add(user.getId(),productId,count);
@@ -66,9 +81,14 @@ public class CartController {
             @ApiImplicitParam(value = "产品id",name = "productId",required = true,paramType = "query"),
             @ApiImplicitParam(value = "数量",name = "count",required = true,paramType = "query"),
     })
-    public ServerResponse update(@ApiIgnore HttpSession session,Integer count,Integer productId) {
-        User user = (User) session.getAttribute(Const.CURRENT_USER);
-        if (user == null) {
+    public ServerResponse update(@ApiIgnore HttpServletRequest request,Integer count,Integer productId) {
+        String loginToken = CookieUtil.readLoginToken(request);
+        if (StringUtils.isEmpty(loginToken)) {
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
+        }
+        String userJsonStr = RedisPoolUtil.get(loginToken);
+        User user = JsonUtil.string2Obj(userJsonStr,User.class);
+        if (user == null){
             return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
         }
         return iCartService.update(user.getId(),productId,count);
@@ -82,9 +102,14 @@ public class CartController {
     @ApiImplicitParams({
             @ApiImplicitParam(value = "产品id列表",name = "productIds",required = true,paramType = "query"),
     })
-    public ServerResponse deleteProduct(@ApiIgnore HttpSession session,String productIds) {
-        User user = (User) session.getAttribute(Const.CURRENT_USER);
-        if (user == null) {
+    public ServerResponse deleteProduct(@ApiIgnore HttpServletRequest request,String productIds) {
+        String loginToken = CookieUtil.readLoginToken(request);
+        if (StringUtils.isEmpty(loginToken)) {
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
+        }
+        String userJsonStr = RedisPoolUtil.get(loginToken);
+        User user = JsonUtil.string2Obj(userJsonStr,User.class);
+        if (user == null){
             return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
         }
         return iCartService.delete(user.getId(),productIds);
@@ -94,9 +119,14 @@ public class CartController {
     @RequestMapping(value = "select_all.do",method = RequestMethod.PUT)
     @ResponseBody
     @ApiOperation("全选")
-    public ServerResponse selectAll(@ApiIgnore HttpSession session) {
-        User user = (User) session.getAttribute(Const.CURRENT_USER);
-        if (user == null) {
+    public ServerResponse selectAll(@ApiIgnore HttpServletRequest request) {
+        String loginToken = CookieUtil.readLoginToken(request);
+        if (StringUtils.isEmpty(loginToken)) {
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
+        }
+        String userJsonStr = RedisPoolUtil.get(loginToken);
+        User user = JsonUtil.string2Obj(userJsonStr,User.class);
+        if (user == null){
             return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
         }
         return iCartService.selectOrUnselect(user.getId(),null,Const.Cart.CHECKED);
@@ -106,9 +136,14 @@ public class CartController {
     @RequestMapping(value = "unselect_all.do",method = RequestMethod.PUT)
     @ResponseBody
     @ApiOperation("全不选")
-    public ServerResponse unSelectAll(@ApiIgnore HttpSession session) {
-        User user = (User) session.getAttribute(Const.CURRENT_USER);
-        if (user == null) {
+    public ServerResponse unSelectAll(@ApiIgnore HttpServletRequest request) {
+        String loginToken = CookieUtil.readLoginToken(request);
+        if (StringUtils.isEmpty(loginToken)) {
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
+        }
+        String userJsonStr = RedisPoolUtil.get(loginToken);
+        User user = JsonUtil.string2Obj(userJsonStr,User.class);
+        if (user == null){
             return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
         }
         return iCartService.selectOrUnselect(user.getId(),null,Const.Cart.UN_CHECKED);
@@ -120,9 +155,14 @@ public class CartController {
     @ApiImplicitParams({
             @ApiImplicitParam(value = "产品id",name = "productId",required = true,paramType = "query"),
     })
-    public ServerResponse select(@ApiIgnore HttpSession session,Integer productId) {
-        User user = (User) session.getAttribute(Const.CURRENT_USER);
-        if (user == null) {
+    public ServerResponse select(@ApiIgnore HttpServletRequest request,Integer productId) {
+        String loginToken = CookieUtil.readLoginToken(request);
+        if (StringUtils.isEmpty(loginToken)) {
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
+        }
+        String userJsonStr = RedisPoolUtil.get(loginToken);
+        User user = JsonUtil.string2Obj(userJsonStr,User.class);
+        if (user == null){
             return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
         }
         return iCartService.selectOrUnselect(user.getId(),productId,Const.Cart.CHECKED);
@@ -135,9 +175,14 @@ public class CartController {
     @ApiImplicitParams({
             @ApiImplicitParam(value = "产品id",name = "productId",required = true,paramType = "query"),
     })
-    public ServerResponse unSelect(@ApiIgnore HttpSession session,Integer productId) {
-        User user = (User) session.getAttribute(Const.CURRENT_USER);
-        if (user == null) {
+    public ServerResponse unSelect(@ApiIgnore HttpServletRequest request,Integer productId) {
+        String loginToken = CookieUtil.readLoginToken(request);
+        if (StringUtils.isEmpty(loginToken)) {
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
+        }
+        String userJsonStr = RedisPoolUtil.get(loginToken);
+        User user = JsonUtil.string2Obj(userJsonStr,User.class);
+        if (user == null){
             return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
         }
         return iCartService.selectOrUnselect(user.getId(),productId,Const.Cart.UN_CHECKED);
@@ -148,10 +193,15 @@ public class CartController {
     @RequestMapping(value = "get_cart_product_count.do",method = RequestMethod.GET)
     @ResponseBody
     @ApiOperation("购物车产品数量")
-    public ServerResponse getCartProductCount(@ApiIgnore HttpSession session) {
-        User user = (User) session.getAttribute(Const.CURRENT_USER);
-        if (user == null) {
-            return ServerResponse.createBySuccess(0);
+    public ServerResponse getCartProductCount(@ApiIgnore HttpServletRequest request) {
+        String loginToken = CookieUtil.readLoginToken(request);
+        if (StringUtils.isEmpty(loginToken)) {
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
+        }
+        String userJsonStr = RedisPoolUtil.get(loginToken);
+        User user = JsonUtil.string2Obj(userJsonStr,User.class);
+        if (user == null){
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEES_LOGIN.getCode(),ResponseCode.NEES_LOGIN.getDesc());
         }
         return iCartService.getCartProductCount(user.getId());
     }
